+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support for async schema fixtures in `schemathesis.pytest.from_fixture`.
 - Generate empty header values during the coverage phase.
 - Correctly handle second CTRL-C when waiting for worker threads.
+- Generate negative values for `minItems` & `maxItems` during the coverage phase.
 
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21